+++ packages/rainbowkit/src/components/RainbowKitProvider/RainbowKitChainContext.tsx
@@ -4,9 +4,9 @@ import { provideRainbowKitChains } from './provideRainbowKitChains';
 
 export interface RainbowKitChain {
   id: number;
+  name?: string;
   iconUrl?: string | (() => Promise<string>) | null;
   iconBackground?: string;
-  name: string;
 }
 
 // This type is a combination of wagmi and RainbowKit chain types to make