+++ packages/rainbowkit/src/wallets/walletConnectors/metaMaskWallet/metaMaskWallet.ts
@@ -60,6 +60,7 @@ function isMetaMask(ethereum?: (typeof window)['ethereum']): boolean {
   if (ethereum.isRabby) return false;
   if (ethereum.isRainbow) return false;
   if (ethereum.isStatus) return false;
+  if (ethereum.isTalisman) return false;
   if (ethereum.isTally) return false;
   if (ethereum.isTokenPocket) return false;
   if (ethereum.isTokenary) return false;
@@ -136,7 +137,7 @@ export const metaMaskWallet = ({
           ? uri
           : isIOS()
           ? // currently broken in MetaMask v6.5.0 https://github.com/MetaMask/metamask-mobile/issues/6457
-            `metamask:///wc?uri=${encodeURIComponent(uri)}`
+            `metamask://wc?uri=${encodeURIComponent(uri)}`
           : `https://metamask.app.link/wc?uri=${encodeURIComponent(uri)}`;
       };